plausibility fix, and removed plausibility for testing

+++ teensy_dash/include/io_settings.hpp
@@ -69,7 +69,7 @@ namespace config {
         constexpr uint16_t MIN_FOR_TORQUE = 0;
         constexpr uint16_t MAX_FOR_TORQUE = MAX - MIN;
         
-        constexpr int8_t ERROR_PLAUSIBILITY = -4;
+        constexpr int ERROR_PLAUSIBILITY = -4;
 
         constexpr uint8_t MAX_ERROR_PERCENT = 10;
         constexpr uint16_t MAX_ERROR_ABS = UPPER_BOUND_APPS_HIGHER * MAX_ERROR_PERCENT/100;
